New models Friend, Review, Comment UserID to BigINteger

+++ src/main/java/com/jof/springmvc/dao/UserDao.java
@@ -2,12 +2,13 @@ package com.jof.springmvc.dao;
 
 import com.jof.springmvc.model.User;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
 public interface UserDao {
 
-    User findById(int id);
+    User findById(BigInteger id);
 
     User findByUsername(String username);
