+++ src/main/java/com/jof/springmvc/service/UserService.java
@@ -2,12 +2,13 @@ package com.jof.springmvc.service;
 
 import com.jof.springmvc.model.User;
 
+import java.math.BigInteger;
 import java.util.List;
 
 
 public interface UserService {
 
-    User findById(int id);
+    User findById(BigInteger id);
 
     User findByUserName(String username);
 
@@ -19,6 +20,6 @@ public interface UserService {
 
     List<User> findAllUsers();
 
-    boolean isUsernameUnique(Integer id, String username);
+    boolean isUsernameUnique(BigInteger id, String username);
 
 }
